wip: start using typescript
