+++ .prettierrc
@@ -1,4 +1,3 @@
 {
-  "semi": false,
-  "parser": "flow"
+  "semi": false
 }
