+++ package.json
@@ -18,7 +18,8 @@
     "prettier": "yarn run lint:eslint --fix",
     "release": "node ./scripts/release.js",
     "release:build": "node ./scripts/build/",
-    "release:test": "node ./scripts/test.js"
+    "release:test": "node ./scripts/test.js",
+    "type-check": "tsc"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -31,7 +32,12 @@
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.1.0",
     "@babel/runtime": "^7.0.0",
+    "@types/babel-types": "^7.0.4",
+    "@types/jest": "^23.3.13",
+    "@types/ramda": "types/npm-ramda#dist",
+    "@typescript-eslint/parser": "^1.2.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
     "babel-jest": "^23.6.0",
@@ -40,14 +46,14 @@
     "enzyme": "^3.5.0",
     "enzyme-adapter-react-16": "^1.3.0",
     "enzyme-to-json": "^3.3.4",
-    "eslint": "^5.4.0",
+    "eslint": "^5.12.1",
     "eslint-plugin-flowtype": "^2.50.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^21.22.0",
-    "eslint-plugin-node": "^7.0.1",
-    "eslint-plugin-prettier": "^2.6.2",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jest": "^22.2.2",
+    "eslint-plugin-node": "^8.0.1",
+    "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-promise": "^4.0.0",
-    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-react": "^7.12.4",
     "filesize": "^3.5.11",
     "flow-bin": "^0.81.0",
     "flow-copy-source": "^2.0.2",
@@ -64,7 +70,7 @@
     "nixt": "^0.5.0",
     "node-emoji": "^1.6.1",
     "ora": "^3.0.0",
-    "prettier": "^1.14.2",
+    "prettier": "^1.16.3",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
     "regenerator-runtime": "^0.12.1",
@@ -80,6 +86,8 @@
     "strip-ansi": "^4.0.0",
     "targz": "^1.0.1",
     "tmp": "^0.0.33",
+    "ts-jest": "^23.10.5",
+    "typescript": "^3.3.1",
     "webpack": "^4.17.1"
   },
   "workspaces": [
