+++ packages/cli/package.json
@@ -59,7 +59,7 @@
     "pofile": "^1.0.11",
     "pseudolocale": "^1.1.0",
     "ramda": "^0.25.0",
-    "typescript": "^2.9.2"
+    "typescript": "^3.3.1"
   },
   "devDependencies": {
     "mockdate": "^2.0.2"
