+++ packages/macro/index.ts
@@ -1,0 +1,2 @@
+import macro from "./src"
+export default macro
