+++ packages/macro/package.json
@@ -2,7 +2,7 @@
   "name": "@lingui/macro",
   "version": "0.0.0-managed-by-release-script",
   "description": "Macro for generating messages in ICU MessageFormat syntax",
-  "main": "index.js",
+  "main": "index.ts",
   "author": {
     "name": "Tomáš Ehrlich",
     "email": "tomas.ehrlich@gmail.com"
