+++ packages/macro/src/macroJsx.ts
@@ -1,0 +1,360 @@
+import * as R from "ramda"
+import ICUMessageFormat from "./icu"
+import { zip } from "./utils"
+import * as babelTypes from "@babel/types"
+
+const pluralRuleRe = /(_[\d\w]+|zero|one|two|few|many|other)/
+const jsx2icuExactChoice = value =>
+  value.replace(/_(\d+)/, "=$1").replace(/_(\w+)/, "$1")
+
+const makeCounter = (index = 0) => () => index++
+
+// replace whitespace before/after newline with single space
+const keepSpaceRe = /\s*(?:\r\n|\r|\n)+\s*/g
+// remove whitespace before/after tag or expression
+const stripAroundTagsRe = /(?:([>}])(?:\r\n|\r|\n)+\s*|(?:\r\n|\r|\n)+\s*(?=[<{]))/g
+
+function normalizeWhitespace(text) {
+  return (
+    text
+      .replace(stripAroundTagsRe, "$1")
+      .replace(keepSpaceRe, " ")
+      // keep escaped newlines
+      .replace(/\\n/, "\n")
+      .replace(/\\s/, " ")
+      .trim()
+  )
+}
+
+export default class MacroJSX {
+  types: typeof babelTypes
+  expressionIndex: () => Number
+  elementIndex: () => Number
+
+  constructor({ types }) {
+    this.types = types
+    this.expressionIndex = makeCounter()
+    this.elementIndex = makeCounter()
+  }
+
+  replacePath = path => {
+    const tokens = this.tokenizeNode(path.node)
+
+    const messageFormat = new ICUMessageFormat()
+    const {
+      message: messageRaw,
+      values,
+      jsxElements
+    } = messageFormat.fromTokens(tokens)
+    const message = normalizeWhitespace(messageRaw)
+
+    const { attributes, id, comment } = this.stripMacroAttributes(path.node)
+
+    if (!id && !message) {
+      return
+    } else if (id && id !== message) {
+      // If `id` prop already exists and generated ID is different,
+      // add it as a `default` prop
+      attributes.push(
+        this.types.jsxAttribute(
+          this.types.jsxIdentifier("id"),
+          this.types.stringLiteral(id)
+        )
+      )
+
+      if (process.env.NODE_ENV !== "production") {
+        if (message) {
+          attributes.push(
+            this.types.jsxAttribute(
+              this.types.jsxIdentifier("defaults"),
+              this.types.stringLiteral(message)
+            )
+          )
+        }
+      }
+    } else {
+      attributes.push(
+        this.types.jsxAttribute(
+          this.types.jsxIdentifier("id"),
+          this.types.stringLiteral(message)
+        )
+      )
+    }
+
+    if (process.env.NODE_ENV !== "production") {
+      if (comment) {
+        attributes.push(
+          this.types.jsxAttribute(
+            this.types.jsxIdentifier("comment"),
+            this.types.stringLiteral(comment)
+          )
+        )
+      }
+    }
+
+    // Parameters for variable substitution
+    const valuesObject = Object.keys(values).map(key =>
+      this.types.objectProperty(this.types.identifier(key), values[key])
+    )
+
+    if (valuesObject.length) {
+      attributes.push(
+        this.types.jsxAttribute(
+          this.types.jsxIdentifier("values"),
+          this.types.jsxExpressionContainer(
+            this.types.objectExpression(valuesObject)
+          )
+        )
+      )
+    }
+
+    // Inline elements
+    if (Object.keys(jsxElements).length) {
+      attributes.push(
+        this.types.jsxAttribute(
+          this.types.jsxIdentifier("components"),
+          this.types.jsxExpressionContainer(
+            this.types.objectExpression(
+              Object.keys(jsxElements).map(key =>
+                this.types.objectProperty(
+                  this.types.identifier(key),
+                  jsxElements[key]
+                )
+              )
+            )
+          )
+        )
+      )
+    }
+
+    path.replaceWith(
+      this.types.jsxElement(
+        this.types.jsxOpeningElement(
+          this.types.jsxIdentifier("Trans"),
+          attributes,
+          /*selfClosing*/ true
+        ),
+        /*closingElement*/ null,
+        /*children*/ [],
+        /*selfClosing*/ true
+      )
+    )
+  }
+
+  stripMacroAttributes = node => {
+    const { attributes } = node.openingElement
+    const attrName = (names, exclude = false) => {
+      const namesRe = new RegExp("^(" + names.join("|") + ")$")
+      return attr =>
+        exclude ? !namesRe.test(attr.name.name) : namesRe.test(attr.name.name)
+    }
+    const id = attributes.filter(attrName(["id"]))[0]
+    const defaults = attributes.filter(attrName(["defaults"]))[0]
+    const comment = attributes.filter(attrName(["comment"]))[0]
+
+    let reserved = ["id", "defaults", "comment"]
+    if (this.isI18nComponent(node)) {
+      // no reserved prop names
+    } else if (this.isChoiceComponent(node)) {
+      reserved = [
+        ...reserved,
+        "_\\w+",
+        "_\\d+",
+        "zero",
+        "one",
+        "two",
+        "few",
+        "many",
+        "other",
+        "value",
+        "offset"
+      ]
+    }
+
+    return {
+      id: id != null ? id.value.value : null,
+      defaults,
+      comment,
+      attributes: attributes.filter(attrName(reserved, true))
+    }
+  }
+
+  tokenizeNode = node => {
+    if (this.isI18nComponent(node)) {
+      // t
+      return this.tokenizeTrans(node)
+    } else if (this.isChoiceComponent(node)) {
+      // plural, select and selectOrdinal
+      return this.tokenizeChoiceComponent(node)
+    } else if (this.types.isJSXElement(node)) {
+      return this.tokenizeElement(node)
+    } else {
+      return this.tokenizeExpression(node)
+    }
+  }
+
+  tokenizeTrans = node => {
+    return R.flatten(
+      node.children.map(child => this.tokenizeChildren(child)).filter(Boolean)
+    )
+  }
+
+  tokenizeChildren = node => {
+    if (this.types.isJSXExpressionContainer(node)) {
+      const exp = node.expression
+
+      if (this.types.isStringLiteral(exp)) {
+        // Escape forced newlines to keep them in message.
+        return {
+          type: "text",
+          value: exp.value.replace(/\n/g, "\\n")
+        }
+      } else if (this.types.isTemplateLiteral(exp)) {
+        const tokenize = R.pipe(
+          // Don't output tokens without text.
+          R.evolve({
+            quasis: R.map((text: babelTypes.TemplateElement) => {
+              // Don't output tokens without text.
+              const value = text.value.raw
+              if (value === "") return null
+
+              return this.tokenizeText(value)
+            }),
+            expressions: R.map(exp =>
+              this.types.isCallExpression(exp)
+                ? this.tokenizeNode(exp)
+                : this.tokenizeExpression(exp)
+            )
+          }),
+          exp => zip(exp.quasis, exp.expressions),
+          R.flatten,
+          R.filter(Boolean)
+        )
+
+        return tokenize(exp)
+      } else if (this.types.isJSXElement(exp)) {
+        return this.tokenizeNode(exp)
+      } else {
+        return this.tokenizeExpression(exp)
+      }
+    } else if (this.types.isJSXElement(node)) {
+      return this.tokenizeNode(node)
+    } else if (this.types.isJSXSpreadChild(node)) {
+      // just do nothing
+    } else if (this.types.isJSXText(node)) {
+      // node.value has HTML entities converted to characters, but we need original
+      // HTML entities.
+      return this.tokenizeText((node as any).extra.raw)
+    } else {
+      // Same as above, but this time node.extra.raw also includes surrounding
+      // quotes. We need to strip them first.
+      return this.tokenizeText(node.extra.raw.replace(/(["'])(.*)\1/, "$2"))
+    }
+  }
+
+  tokenizeChoiceComponent = node => {
+    const element = node.openingElement
+    const format = element.name.name.toLowerCase()
+    const props = element.attributes
+
+    const token = {
+      type: "arg",
+      format,
+      name: null,
+      value: undefined,
+      options: {
+        offset: undefined
+      }
+    }
+
+    for (const attr of props) {
+      const name = attr.name.name
+
+      if (name === "value") {
+        const exp = this.types.isLiteral(attr.value)
+          ? attr.value
+          : attr.value.expression
+        token.name = this.expressionToArgument(exp)
+        token.value = exp
+      } else if (format !== "select" && name === "offset") {
+        // offset is static parameter, so it must be either string or number
+        token.options.offset = this.types.isStringLiteral(attr.value)
+          ? attr.value.value
+          : attr.value.expression.value
+      } else if (pluralRuleRe.test(name)) {
+        let value
+        if (this.types.isStringLiteral(attr.value)) {
+          value = attr.value.extra.raw.replace(/(["'])(.*)\1/, "$2")
+        } else {
+          value = this.tokenizeChildren(attr.value)
+        }
+        token.options[jsx2icuExactChoice(name)] = value
+      }
+    }
+
+    return token
+  }
+
+  tokenizeElement = node => {
+    // !!! Important: Calculate element index before traversing children.
+    // That way outside elements are numbered before inner elements. (...and it looks pretty).
+    const name = this.elementIndex()
+    const children = node.children
+      .map(child => this.tokenizeChildren(child))
+      .filter(Boolean)
+
+    node.children = []
+    node.openingElement.selfClosing = true
+
+    return {
+      type: "element",
+      name,
+      value: node,
+      children
+    }
+  }
+
+  tokenizeExpression = node => {
+    return {
+      type: "arg",
+      name: this.expressionToArgument(node),
+      value: node
+    }
+  }
+
+  tokenizeText = value => {
+    return {
+      type: "text",
+      value
+    }
+  }
+
+  expressionToArgument = exp => {
+    return this.types.isIdentifier(exp) ? exp.name : this.expressionIndex()
+  }
+
+  /**
+   * Custom matchers
+   */
+
+  isIdentifier = (node, name) => {
+    return this.types.isIdentifier(node, { name })
+  }
+
+  isI18nComponent = (node, name = "Trans") => {
+    return (
+      this.types.isJSXElement(node) &&
+      this.types.isJSXIdentifier(node.openingElement.name, {
+        name
+      })
+    )
+  }
+
+  isChoiceComponent = node => {
+    return (
+      this.isI18nComponent(node, "Plural") ||
+      this.isI18nComponent(node, "Select") ||
+      this.isI18nComponent(node, "SelectOrdinal")
+    )
+  }
+}
