+++ packages/macro/src/utils.ts
