+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "jsx": "react",
+    "noEmit": true,
+    "pretty": true,
+    "esModuleInterop": true,
+    "target": "es2018"
+  }
+}
